fix(hosted): execute opaque runtime tools from workspace
