+++ private/hosted-cli-candidate/runtime-config.js
@@ -21,6 +21,7 @@ const RESERVED_ENVIRONMENT = new Set([
   'TMPDIR',
   'ZEROSHOT_HOSTED_BASE_REVISION',
   'ZEROSHOT_HOSTED_EXECUTABLE',
+  'ZEROSHOT_HOSTED_EXEC_ROOT',
   'ZEROSHOT_HOSTED_MODEL',
   'ZEROSHOT_HOSTED_PROVIDER',
   'ZEROSHOT_HOSTED_REPOSITORY',
