+++ zeroshot-rust/hosted-node/hosted-config.js
@@ -5,6 +5,7 @@ const fs = require('node:fs');
 const REPOSITORY_ENV = 'ZEROSHOT_HOSTED_REPOSITORY';
 const BASE_REVISION_ENV = 'ZEROSHOT_HOSTED_BASE_REVISION';
 const EXECUTABLE_ENV = 'ZEROSHOT_HOSTED_EXECUTABLE';
+const EXECUTABLE_ROOT_ENV = 'ZEROSHOT_HOSTED_EXEC_ROOT';
 const PROVIDER_ENV = 'ZEROSHOT_HOSTED_PROVIDER';
 const MODEL_ENV = 'ZEROSHOT_HOSTED_MODEL';
 const SETTINGS_ENV = 'ZEROSHOT_SETTINGS_FILE';
@@ -26,6 +27,7 @@ const CONTROL_ENVIRONMENT = new Set([
   'TMPDIR',
   BASE_REVISION_ENV,
   EXECUTABLE_ENV,
+  EXECUTABLE_ROOT_ENV,
   MODEL_ENV,
   PROVIDER_ENV,
   REPOSITORY_ENV,
